Test Controller

+++ www/index.php
@@ -1,27 +1,26 @@
 <?php
 include 'Controller/index_controller.php';
 
+$ctrl = new Controller();
 
+if (!isset($_GET["action"])) {
+    echo $ctrl->create();
+} else {
+    switch ($_GET["action"]) {
+        case "create":
+            echo $ctrl->create();
+            break;
 
-$ctrl = Controller();
+        case "read":
+            echo $ctrl->read();
+            break;
 
-switch ($_GET["action"]) {
-    case "create":
-        echo $ctrl->create();
-        break;
+        case "update":
+            echo $ctrl->update();
+            break;
 
-    case "read":
-        echo $ctrl->read();
-        break;
-
-    case "update":
-        echo $ctrl->update();
-        break;
-
-    case "delete":
-        echo $ctrl->delete();
-        break;
-
-    default:
-        echo $ctrl->create();
+        case "delete":
+            echo $ctrl->delete();
+            break;
+    }
 }
